Optimized code completion match algorithm & made it easier to re-use. The navigate to command now uses this algorithm for filtering.
A large amount of time of the navigate to dialog item list was taken by the filter - using the greedy algorithm from the code completion code makes this feature much faster & it's now more consistent with the code completion. However the old algorithm is still in there and reverting this is a small change in the NavigateToDialog - just uncomment it. When all users are satisfied with the code completion algorithm the code should be taken out.