Fix serialization issue.
When serializing a property, use the property type as base type for
serialization, which means that the type name will not be included
in the serialization. Using System.Object as base type is not
necessary because the type is provided when retrieving the value of
the property. Also, using System.Object as base type does not work
for primitive types.