[Debugger] Added an sdb testing debug session type
New hidden sdb session type that can run an arbitrary command
then initiate a listen or connect sdb session to an arbitrary
address. Not suitable for general use yet, but very useful
for testing sdb and for debugging embedded targets.

To enable it set the env var MONODEVELOP_SDB_TEST=1.