Fix debugger crash on Windows w/redirected console
Windows socket behaviour seems to be slightly different from
Linux/Mac, and Mono.Debugger.Soft was trying to disconnect a socket
that that doesn't get connected, causing an exception at launch.

For some reason it did check the main debugger socket before trying
to disconnect it, but not the console socket, so only apps with
redirected console output would crash.