This branch contains the changes necessary to integrate with Windows 7's JumpLists found by right click the TaskBar entry and in the StartMenu.
In order to integrate with JumpLists there are several registry entries which need to be added. The registry entries are in HKCR and identify which file extensions MonoDevelop can handle and the command to execute MonoDevelop for a specific file. Ideally these registry entries should be done as part of setup, however JumpList.cs will do the necessary things if the process has permission to access the registry.

This change uses the Windows API Code Pack to interact with Windows. JumpList.cs is based on a sample provided with the code pack.

The list of file extensions for files is obtained from the same list of file extensions displayed in the OpenFile dialog. Thus if a file is added there, it will be reflected in the registry and can appear in the JumpList.

In order to set the JumpList the HWND for the window is required. The extension point current exposed to inject start-up code is '/MonoDevelop/Ide/StartupHandlers' however I was not able to reliably obtain the correct HWND. It is either because CommandService.RootWindow is changed during the IdeStartup, initialization timing, or because of timing issues with which window is registered with the system as the main window. One solution that appears to solve the issue is to use /MonoDevelop/Ide/StartupHandlers and then use a delayed callback to update the JumpList some time later. I gave up on this approach as it isn't deterministic and is error prone without a long delay between startup and jumplist registration. (When a JumpList is added to the window the taskbar item changes. It's a weird user experience if this doesn't happen immediately at startup.) To address these timing issues I added a new ExtensionPoint '/MonoDevelop/Ide/InitCompleteHandlers' which is executed just like StartupHandlers except done after MonoDevelop is fully loaded.