[Mac] Merge entire install prefix into app bundle
This allows us to use a newer Mono.Addins than is installed on
the system.

Note that the entire prefix is merged into the app bundle. Previous
changes allow MD to automatically pick up everything installed in
the prefix at configure time and build time, and with "make run".
Launch scripts now pick up the prefix within the app bundle.