[Ide] Zoom tree pads in a reversible way
Copied the logic from the text editor widget so that
the zooming in and zooming out steps are the same size.
That means that zooming the same number of steps out
and back in no longer leaves you with a zoom value 
that's almost but not quite 1, and fuzzy rendering.