Fix issue when closing the main window in Mac
The problem is that the HasTopLevelFocus property is not being set
when restoring the main window, then MD thinks that it doesn't have
the focus and all commands are disabled. This is a bug in GTK+. The
workaround is to create a dummy top level window and manually generate
a focus change.