[MacDev] Don't re-use Xcode project directories
Xcode keeps some kind of internal lock on project files while 
it's running and gets extremely unhappy if a new or changed
project is in the same location as a previously opened one.

To work around this we increment a subdirectory name and 
use that, and do some careful bookkeeping to reduce the 
unnecessary I/O overhead that this adds.