[CrashMonitor] Begin implementing a basic crash monitor for MonoDevelop
This application will run as a separate process which will monitor for native
crashes in MonoDevelop. Initially it will work on MacOS and will check for
crash logs written by the MacOS Crash Reporter tool. This will eventually
be expanded to other platforms as appropriate. The crash reports will be
uploaded to xamarin for analysis.