[MacCore] Give a useful warning if XCode becomes unresponsive.
Should XCode crash at just the wrong time it is possible for the user
to get an exception from exceuting the AppleScript used to interop with
it. Try to handle this, and any other case where we have an exception
when trying to get XCode to save any unsaved modifications, by giving a
moderately useful message.