[MacCore] Sort items properly in the xcode solution
The original attempt to sort the items when adding them to the generated
xcode solution sorted them by calling ToString and comparing that. This
is useless as ToString is typically prefixed with a random ID and so you
will always end up with random sorting. Fix this by putting some smarts
into the sorting so it can handle (at the least) files and groups.