[Ide] Allow ProjectFileSelector to choose files in subdirs
Previously if we had the directory structure 'Dir1/Dir2' inside the
project, the file selector would incorrectly generate the path
'{project_root}/Dir2' when you navigated to Dir2. This is because it
was appending the name of the current node to the root project path.

Resolve this by storing the name of the dir and the FilePath corresponding
to the dir in the TreeStore so we can trivially append the correct path
to the project root.