A fix for a compiler and path error.
The error has been observed while compiling on Windows 7.
Due to a path not wrapped in double quotes,
the compiling fails. This is likely due to spaces
occuring in the path. The fix ensures that paths
are wrapped in quotes when compiling.
Wrapping is not used in LanguageService.fs,
since the IntelliSense may fail to load assemblies
if it gets a wrapped path.
The fix has been tested on Debian Wheezy and Windows 7.