[MacDev] Moved Objc->Cli type resolution back out of the header parser.
Moved the Objc->Cli type resolution logic back out of the parser since
we'll need to move type resolution to a later stage in the near future.

Also cleaned up the type resolution method into 2 logical parts:
 - Cli -> Objc
 - Objc -> Cli

This makes the code a bit more readable and flexible as we'll need to do
more complex resolution in the Objc->Cli case in the near future. As a
first step, this method also now returns the number of unresolved types
so that callers can have a better idea of where things stand. This will
be important going forward if/when the user ever creates types that are
inter-dependant.