[Core] Disallow spaces in project names
IPhone projects with embedded spaces in their names were creating xib files
where the class name embedded in the xib had a space. This resulted
in breakage when loading the xib in xcode as the embedded name did not
match the name of the actual obj-c class. The simplest fix is to
simply disallow spaces in project names as this kind of issue could
bite us in many places otherwise. Fixes bug #933.