[MacDev] Break type resolution into 2 steps
We can get away easily with a 2-pass type resolution if we assign FQN
type names to newly added user-defined types in the first pass and do
a standard type-resolution pass on the second.

Also add newly added project files to the sync context's LastSyncTime
table so that it knows about them.