[Core] Refactor the UI out of the project migration handlers
When loading a project, an IProjectLoadProgressHandler monitor should be
supplied. This has an additional method which asks the user if the migration
should happen and if a backup should be made. The answer to the prompt is
applied to every project which needs migrating so the user only ever gets
prompted once. Fixes #955.