[Xcode] Handle NeedsSyncBack/NeedsSyncOut better with weird modification times
When syncing to/from Xcode, we should check that the modification times match
the cached one exactly rather than checking if one is greater than the other.
This way we can handle strange cases where the user has a file which is
created/modified in the future like in bug #1137.