[Ide] Better fix for replacing with an empty string
Redo the change in 96c55e so we can properly detect a replace with
empty string versus a plain old file. Whenever the 'Replace' button
is clicked we ensure a non-null string is passed as the replace
pattern. If 'Find' is clicked, we pass in null. This means we need
to use replacePattern == null checks instead of string.IsNullOrEmpty
checks.