[MacDev] All the image chooser to select multiple image formats
Improve the ImageChooser to allow the user to specify the image formats
which should be allowed. For example sometimes the user only wants png
files, sometimes jpg, sometimes any format will do. By default it allows
png only.

When a file is selected, the chooser then verifies that it is of the
correct type and also if a size has been specified, it verifies the
image is of the correct size.