[Core] Move log file redirection/creation to LoggingService
The LoggingService is now in charge of deleting old logs, creating new
logs and redirecting stdout/stderr. The session guid is simply
DateTime.Now.Ticks.ToString. This should be unique enough to never have
duplicates but it also has the benefit that the log files will have a
natural sorting order so getting the newest log file to attach to a bug
report should be trivial for the user.