[Core] Fix the implement of DragNotebook.FindTabAtPosition
Offscreen widgets still have an allocation > 0 so in order to select
the correct widget to display the context menu on the correct tab page
we need to walk up/down the tabcontrol from the currently selected
page until we hit the first widget satisfying the x/y coordinates.

Fixes a bug where scrolling the tab view to the right to put tabs
14-20 in view and right clicking on tab 14 would actually select
tab 0 and display the popup for that tab.