[Ide] Toolbar buttons should have tooltips
If toolbar buttons have an icon, the label is not displayed, but
the  description is shown as a tooltip. However, if the description
is not set, it's not always clear what the button does. In such
cases we now fall back to displaying the label as a tooltip.

Fixes Bug 526 - Tooltips not available on "Breakpoints" window