[Ide] Only copy project files when dragging/dropping nodes
We cannot afford to copy files which are not in the project without
getting deeper integration between the version control system and
MonoDevelop. If we blindly copy everything we will end up copying
the .svn directory which results in it and all its contents being
added as project files too. We also try to add the contents of the
.svn directory to the VCS system too. Blindly copying everything
could also cause problems with other VC systems which have special
directories.