[Ide] Only add files to the VCS, not directories
If i drag/drop a directory from one project to another the SVN addin
blows up because it tries to add a directory as if it were a file. There
doesn't appear to be an easy way to distinguish between files and
directories in the addin so the simplest thing to do is to ignore all
directories. The result is that empty directories will not be added to
the VCS if you drag/drop a directory which contains an empty directory,
but everything else will work just fine. Fixes part of bug #2038.