[TextEditor] Fixes rendering on gtk-win32 2.24.8
With GTK+ 2.24.8 on Windows, there was an issue where children of
the TextEditorContainer widget would not get map/unmap events. This
caused the text editor widgets in the main notebook to turn black
the second and subsequent times they were mapped by switching the
notebook tabs.

This seems to be because the TextEditorContainer had a GdkWindow,
but was marked as NoWindow. Since the GdkWindow appears to be
completely redundant (and possibly leaky), I removed it.