[Core] FileService.Rename should not emit FileRemove and FileAdd events
If we rename a file on a case insensitive filesystem when the only
change in the filename is a casing change, we can break if a Rename
command emits both an Add and Remove event. The reason for this is
that the FilePaths for the old path and new path will compare the same
and this will result in the file closing if the user has it open.
Fixes the remainder of bug #2437.