[MacDev] Fix an issue where we'd constantly sync files to xcode
We check if content files have been modified by comparing the mtime
of the xcode version with the mtime of the monodevelop original. As
such we should ensure that when we sync the file out, it must have
the same mtime as the original file.

Fixes a bug where content files would permanently return true for
'needs sync' which in turn caused us to execute lots of applescript
and also copy the files again every time a sync out is triggered,
such as when a file is saved. A similar fix was required for regular
source files too.