[Core] Ensure that MonoDevelop.log always points to the latest log file
On Mac/Linux we use a symlink to ensure that MonoDevelop.log always
points to the most recently created log file. On windows we always
write the latest log file to 'MonoDevelop.log' or MonoDevelop-{0}.log
and then rename them to the timestamped format the next time MonoDevelop
is launched.