[Mono.Debugger.Soft] Fix debugging csc-compiled code
When csc is run normally (i.e. not the in-process VS compiler) it lowercases
the drivenames of the source files referred to by the pdb file. However, drivenames
are canonically uppercased.

The new type load event code reduces wire traffic setting a filter on the type
load event request to restrict the events to the types in certain files. This
filename string matching is done in the runtime in a case sensitive way, so
cannot handle the inconsistency in the csc-compiled symbols. As a result,
breakpoints cannot be inserted.

Until the runtime support case-insensitive filename string matching, we work
around this issue by requesting type load events for both casing variants of
the filename.

Fixes BXC #2196 - [MD/Win] Debugging doesn't stop at breakpoints.