[Core] Add support in core for saving/retrieving passwords
Add an IPasswordProvider interface and a PasswordService which allows
monodevelop to store/retrieve passwords in a platform dependent way.

[MacDev] Add a password provider implementation for the mac Keychain

We can now save/restore passwords in a secure way using the keychain.
The git addin now uses the PasswordService to store/retrieve passwords
for git repos which require them.

Partial fix for bug #2737.