[VersionControl] Properly support passphrases and invalid passwords/passphrases
We can now store passphrases in the keychain so passphrase protected ssh keys
loaded by ngit can be handled properly. We also handle the case where passwords
are invalid. If we successfully retrieve a password and are queried again for
the password within a short amount of time, we will bypass the password store
and instead query the user again for a password.