[Ide] Fix #2780 Drag-drop of file in solution pad does nothing
This was mostly a cosmetic issue, but it was misleading and confusing.
The ExtensibleTreeView was handling the hover event and when it could
determine the drop as valid, was letting it fall through to the treeview's
default handler, which was setting the drop marker. Unfortunately, it
didn't fall through in the cases where the drop was not valid, so the
marker was never unset.

This meant that the drop marker would display in cases when the drop
was not actually valid, and dropping would do nothing. The specific case was
at the edges of an item, which would be a "before" or "after" drop, which was
invalid for project folders. Because of the bug, it displayed as an "into"
drop marker, which was incorrect.

This fix involves always calling the default handler, but explitly unsetting
the marker afterwards if the drop is not valid. This means that we will also
get better hover-to-open and autoscrolling.