[Docking] Fix a case where a widget could be parented twice
The DockItem widgets can be parented into a floating window, the DockFrame's
DockContainer, or an AutoHideDock. In the latter two cases, it's possible
that a race could cause the widget to be added to its new parent before it's
removed from its old parent.

May fix bug 1445 - MD crashes with SIGABRT occasionally when stopping debugging