[Core] When a SolutionEntityItem is renamed, ensure we update the last save time
If we rename the solution file, the on-disk file changes location so we need
to update the project file tracker so that it does not return DateTime.MinValue
when we query the last save time. We also need to ensure that we rename the file
on disk before we rename the item in memory as renaming the in-memory
representation causes the filesystem to be probed for a file with that name.

Fixes bug #2885.