[VersionControl] Add a faster way of generating diffs for the status view
The old code used to walk the entire solution to gather all the
VersionInfo objects for changed files. It would then walk the entire
filesystem a second time as soon as you expanded one of hte diffs so you
could view it. We can completely remove the second walk by simply re-using
the versioninfo objects we gathered during the first pass, typically only
a few seconds previously. This means that diffs show up near instantly on
large projects instead of having a large delay as the entire filesystem is
probed again.