[Mono.TextEditor] Realized does not imply that size has been allocated.
Bug #2960 seems to be caused by the assumption that IsRealized will only
be true after the window size has been set. This is not the case.

When CenterTo() is called before the size has been set, it tries to
rely on width and height values (and adjustment page sizes) that
have not yet been set appropriately for the window.