[Ide] Exporting Projects doesn't make sense, only export Solutions
Exporting Projects can't reliably work because it may reference other projects.
This was the cause of the crash in bug #2979.

I modified the code to walk up to the parent Solution when a user tries to
export a project, and then export the entire solution. This was also the
behavior that mhutch had expected when he tried to export a project.