[Ide] Handle re-adding the same file in the same place
If the user tries to add a file to their project which is already
in the project in the same place, silently ignore it instead of
prompting them with a dialog every time.

Fixes the case where the user might select all files in a directory
and attempt to add them to a project in order to trivially add all
the new ones.