[Svn] A non empty diff does not mean we are guaranteed to get a valid DiffInfo
If we succesfully generate a diff from a local path it does not mean we will be able
to generate a valid DiffInfo from it. If a new directory is added to version control
the bug mentioned in #3178 will be triggered causing an IndexOutOfRange exception
because we cannot generate a valid diff for the directory (as expected).