[Ide] Fix races in ProjectOperations.Rebuild
Because it simply chained the clean and build commands together,
and they both dispatch some calls onto the GUI thread, they
didn't seem to be using the monitor correctly. Also, it was
not consistently dispatching the build events on the GUI thread.

This fixes an issue where the build output pad would get
more indented each time that the Rebuild command was run.
It also make it so that the build output pad contains both
the clean output and the build output.