[Ide] Always close all open documents when closing the workspace
Remove the check to see if Document.Project is non-null before closing
a document. When the workspace is being closed it would be expected that
all open documents are closed.

This resolves bug #3172 which was caused by 'Project' always being null
for the version control subviews and will also ensure any other custom
IViewContents with null Projects are closed when the workspace closes.