[VersionControl] Handle reverting new directories which contain files in SVN
If a directory with containing files has just been added to a project under
SVN control and the user then tries to revert that change, we must be careful
to specify the correct order when passing the paths to libsvn. If we pass the
directory first, libsvn will revert the directory and all containing files.
Then ocne it tries to process the file, it will throw an error because it
has already implicitly reverted the file.

Fixes bug #3176.