[Ide] Prevent an exception from bubbling through GTK
It's possible for GetPath to throw an exception if the filenames are
empty or null so just place it inside the try/catch to prevent that
exception from bubbling through GTK. Fixes a stacktrace found through
some crash logs. I do not know what the root cause of the issue is though.