[Ide] Rewrite the logic to find the completion cache to reduce disk access
We used to indiscriminately read all the subdirectories in the cache directory
in order to pick out the correct completion cache file. For solutions like
MonoDevelop where we have dozens of projects this meant that we had to scan
the filesystem and parse 1000's of xml files in order to load 80 projects
worth of cache data.

Rewrite the logic to attempt to pick out the correct path first so that
in the common case we directly open the exact file we need instead of
linearly scanning. Reduces the loading time for the MonoDevelop workspace
by approximately 10% on my desktop.