[CSharpBinding] Emit generic constraints correctly
When emitting generic constraints we need to ignore System.ValueType
and System.Object from mscorlib when figuring out if the user has
specified a type constraint. These values are implicitly added when
calling GetDirectBaseTypes and should never be used in constraints.

Also ensure that type constraints are not emitted for explicit interface
implementations as it does not compile.

Fixes bug #3314.