[VersionControl] Fix a null reference exception and speed up switching between VC subviews
VersionControlDocumentInfo is basically just a class to store the history for a particular
file which is under version control. If this object is shared between all the version
control views, then we would only have to query the repository once and then all the subviews
would have access to the history. As such, refactor the version control addin to use the
IFastVersionControlViewHandler interface where possible so that version control addins
which support creating their views from a single (shared) VersionControlDocumentInfo can do
so.

This also fixes bug #3228 which was caused by the BlameView opening the LogView without 'Start'
being called on the VersionControlDocumentInfo stored by the LogView. This resulted in a null
reference exception.