[VersionControl] (GIT) Speed up getting the status of a file/directory
Remove a bunch of custom logic which was used to generate the status report
for a checkout and use the built-in logic instead. This greatly speeds up
the process as the original code processed every file on disk whereas the
new code can filter out entire subtrees easily and aborts scanning as soon
as it is guaranteed that no further files will satisfy the filter (if there
is one).