[LogReporting] Add the ability to log an error which does not require user interaction
There are some exceptions which could happen very commonly (every time a document is
opened) which we would like feedback about but would not like to prompt the user about
every single time it happens. In this scenario, we log the exception only if they have
opted in to error reporting.