[VersionControl] Attempt to mitigate the exceptions when VersionInfo cannot be calculated
If an exception is thrown while calling GetVersionInfo on a file (typically when attaching
the version control view tabs), we should catch and hide the error. If we do not, the user
will see an error dialog every single time they switch active documents or open a new
document, thus making MonoDevelop unusable.