[Mono.Debugging.Soft] Rename vm.Dispose to Detach
It's not actually a Dispose method, the name was carried over from JDWP.

Stop misusing it in MD. This means Connection can no longer be disposed
to force disconnection. Add a ForceDisconnect method that MD can use
instead.

This is an ABI break but the only known users of the method were using
it incorrectly due to the confusing name, so it's probably for the best.