[MacDev] Refactor the PListScheme to better support subvalues
We now support PList keys with arbitrary depth. I.e. you can have a key
which contains 5 values, each of which contain 5 values, etc. This allows
us to give proper 'code completion' when editing known keys in the plist
editor. Also add features like 'required' so we know when a value must
be created for the key to be valid