[Ide] #1303 - Exception from recent files when opening file
When adding items to the recent files list, the RecentFileStorageClass
was performing removal of the old item and insertion of the new item
as separate operations. The removal would trigger other instances of MD
to read the changed file, and there was a high chance that their lock
on the file would prevent the addition. The class would retry reads
but not writes. The net effect was that items would vanish from the list.

This is fixed by performing all read/remove/add/write operations with
a single file handle.

NOTE: this was only an issue on Windows, which implies either that the
FSW or the file locking were not working reliably on Mac/Linux, or
simply that the race didn't happen due to differences in speed of
operations.