[MacDev] Split up PListScheme.AvailableValues into two methods
There are two distinct use cases for this method. The first use case
is when we are editing the value of a PObject such as a PNumber of PString.
In this scenario we need to know all the valid values the PObject can take.

The second scenario is when we are adding a new key to the root PObject or
we are adding a subkey to an existing key/subkey. Here we need a list of
all the available keys which have not been used.

Tests have been added for both new methods and the plist editor has been
modified to account for this change. This cleans up a lot of code where
we had to get the parent of a pobject in some cases but not others and
makes it easier to understand which method should be called.