[Build] Include MonoMac.dll as part of monodevelops regular build process
This commit contains a number of changes:

1) MonoMac and MacCore are included as git submodules. The regular readonly
git url is used so that external contributors can check the submodules out.
They are stored in external/monomac and external/maccore.

2) MonoMac.dll is built by MacPlatform as this is the first place we require
it.

3) All projects have been updated to use the assembly built in external/monomac

4) The MacCrashLogger project has been removed from git as the application does
absolutely nothing. All logic and classes were moved to monodevelop itself.

5) MonoMac and MacCore have been removed from the version-checks directory as
they are now integrated into the regular monodevelop build.