Added workaround method for getting the current key modifiers
Gtk.Global.GetCurrentEventState only works when run inside a handler of a gtk event.
On mac, the global menu is not implemented in GTK, so when a menu command is
run, there is no current GTK event, so GetCurrentEventState doesn't return
the correct result. I added a workaround implementation which uses NSEvent when
running in mac, which will always return the correct result, no matter if we are handling
a GTK or a Cocoa event.