[Core] Fix a bug and a speed issue in DotNetProject.PopulateSupportFileList
Firstly we were using an initializer on a ThreadStatic variable. This is
incorrect (as per documentation) as the initializer is only run *once*,
not once per thread. Therefore the 'main' thread would view the variable
as having a default value of -1 and everything else would view it as
having a value of 0. Removing the initializer fixes the LocalCopy unit
test.

We now also use a hashset to track which projects have been walked when
calculating the list of support files. This greatly speeds up the case
where there are dozens of projects in a solution and they all reference
each other.