[NUnit] Correctly pre-load the NUnit assemblies correctly for the runner
We needed to split out the assembly loading from the Initialise call. The
reason for this was that when that method was jitted we needed to load
nunit.core.interfaces.dll as the method invoked a function from that
assembly. Mono then looked for that assembly in the appdomains base
directory so if the assembly under test had local copy of that assembly
 we would load it instead of the one in monodevelops copy.

Fixes bug #5716