[Core] Use better detection to rebuild resx files
There are two issues. Firstly we were parsing .resx files to figure
out if they reference any external files and then using those external
files to decide if the application needed to be rebuilt. This is quite
slow and is also different to how Visual Studio behaves. It merely does
a filestamp check on the resx file. This is what we do now.

Secondly, if we have a resx file in a project which is built using Xbuild,
we need to check in the intermediate output directory to find the built
.resources file to check whether or not it is stale. By checking there,
Mono for Android projects no longer think they are always stale if they
contain a .resx file.

This is not a 100% accurate fix unfortunately as we do not respect the
MBuild variable for the intermediate directory as defined in the csproj
file. It's good enough for most cases though.