[Build] Work around an xbuild issue which causes unnecessary rebuilds
Xbuild does not handle local copy correctly which results in unnecessary
rebuilds. The bug appears to be that if Foo.dll is considered to not
need a rebuild, any references which are specified as 'Local Copy' are
copied anyway. The result is that the mtime on those changes which
somehow triggers a rebuild for other assemblies/projects depending on
Foo.dll.