Properly load, reuse and refresh msbuild builders
The msbuild builder process now uses a single msbuild engine instance
for all projects it loads, so projects are now loaded only once, and builds
are faster. Also, when running a target we now provide the required
configuration for the project built and all the referenced projects, based
on the solution configuration mappings.