[Core] Correctly strip out trailing path separator chars
When unix paths are passed to this function on windows, the TrimEnd
will not remove unix-style trailing separator chars. However if we invoke
the TrimEnd after the call to Path.GetFullPath, the unix style separators
will have been converted to windows style ones and so the call will
do what is expected.

Some minor fixups to the other tests to make them platform agnostic
so they pass on Windows too.