[MacDev] Correctly render the values for PListValueObjects
If 'Show Descriptions' is enabled in the plist editor we should only
render the Description if the actual value of the PObject is the same
as the Description or Identifier for that item.

This means that for known keys like UISupportedOrientations we will
render '1' as 'iPhone' (or whatever) but for keys with no specified
value like 'Document Type Name', we render what the user actually
types instead of rendering the literal string 'Document Type Name'.

This issue could be triggered by editing the item inside the specialised
editor for Document Types.

Conflicts:

	main/external/cecil