[VersionControl] Disable updating the solution panel on FocusIn
Attempting to refresh all the nodes on focus-in is too slow when large
numbers of nodes are open. Another issue is that once a node is opened,
it is always updated even if it is collapsed later on, so as time goes
on and more nodes are expanded and collapsed, the refresh takes longer
and longer.

Conflicts:

	main/external/cecil