[gtk] Fix container leak fix on subclasses
The container leak fix didn't work correctly with subclasses of
managed containers - the fix needed to be registered on those
GTypes too.

When this was fixed, the callback needed to be altered to only
handle the exact type for which it was registered, else it would
cause stack overflows via the subclass's callback.