Fix allocation issue in Popopver window
The old implementation did a reallocation of the child of
the popover in order to add the required padding. This is
may end causing infinite reallocation calls in some cases,
such as bug #6911. The padding is now applied using
an Alignment widget.
This fixes 6911 - Regression: MD hangs when hover window overflows screen size