Updates for MD 3.0.4
An API changed in 3.0.4

Also, in Mono 3.0.4, this change
https://github.com/mono/monodevelop/commit/4a95275cf7d0418250d87844550934b849b1153d
appears to have busted all tooltips in editor extensions?  The call to
CreateTooltip in

main/src/addins/MonoDevelop.SourceEditor2/MonoDevelop.SourceEditor/LanguageItemWindow.cs

was commented out in that change, but this means CreateTooltip is never
called and no tooltips are ever created. So instead we just create the
Gtk window ourselves here.