[MSBuild] Attempt to fix project configurations when building with MSBuild
This is a workaround for bug #7141. The way we invoke MSBuild results in the
global properties for 'Configuration' and 'Platform' being cleared so if
a project depends on another project, we end up using the wrong configuration.
For example if we set the build to 'release' then we would try to reference
files from dependent projects based on that projects 'debug' output.