[Core] Fix the invocation of 'al' on Windows
We need to call GetToolPath and pass 'al' in order to get the full path
to the binary so it can be invoked on windows or Mono as appropriate.
Also 'al2' is mono specific, so the handling of al versus al2 has been
moved to the mono framework backend.

Fixes bug #7296.