[Core] Add better heuristics for matching configurations
When a new project is added to a solution, we now use better heuristics
to figure out which project configuration should be used for every
solution level configuration. For example if there's a solution
level configuration called 'Foo|Any CPU', which does not emit the 'DEBUG'
symbol, we should choose a project configuration which also does not
emit the 'DEBUG' symbol.

Fixes bug #6169.