[Ide] Discard glib messages if we're being spammed by thousands of them
If we do something wrong with Gtk, sometimes we get the exact same message
tens of thousands of times which balloons our log files to tens of gigs.
We can fix this by simply discarding messages for the rest of the session
once we've written a certain number of bytes.

For now I choose 1MB of messages as the default limit as this is large
enough to be sure to capture all the info we need but small enough to
allow the user to send us the entire log file.