[Web References] Add configuration dialog for collection types.
This adds a 'Configure' context menu item to launch a dialog
where the user can configure the list and dictionary collection
types.  The dialog is also launched when creating a new Web Reference.

This requires MonoDevelop to run with mono/master commit ea570f3 or
later; there is a runtime check which automatically hides the dialog
if the runtime is too old.