[Ide] Ensure that searching is always cancelled ASAP
If the search window is destroyed, ensure we cancel the current search
if it exists. We also need to ensure that the sorting code is cancellable
too as large solutions can spend a significant amount of time sorting
their results when only 1 or 2 characters have been entered into the
search entry.