[Ide] Fix a thread safety issue in CommandSearchCategory
This search is not thread safe so we need to ensure we run
it in the GUI thread to avoid deadlocks and/or crashes. The
issue is that check if a command is active requires accessing
gtk and it's possible for the background thread to be working
on this while the main Gtk performs an Update on all commands.
We then have two threads accessing the same objects simultaenously.