[Ide] Correctly handle multiple different OutputPaths in the project config editor
We need to handle the case where the user has multiple output paths defined for
their build configurations. In this scenario if you select the 'All Configurations'
item in the dropdown, it will have empty text. We need to ensure that immediately
changing that dialog to another configuration, like 'Debug', does not trigger
an error dialog due to the path being 'invalid'.

Due to the lack of proper save notifications, there is no cleaner way of handling
this as saves are implicit and are performed as soon as the combobox changes.

Fixes bug #7897