[VersionControl] Massively speed up reloading a project
If we are doing a simple project reload, we can skip
this method completely as only fresh projects will have
files we must add to git.

I also greatly sped up the processing of files by changing
the code to do a single version control invocation to
retrieve the status for the entire subdirectory as compared
to retrieving the status for every file individually.