Fix: FSharpBinding project does not build correctly with xbuild because resources are named incorrectly
The previous MonoDevelop binding did its own build system by cracking
the project files. This is now slowly being phased out and replaced by
xbuild.

The problem is, we cant build this add in with xbuild without a small
change. The old way of doing things gave a resource in a subdirectory
such as

abc/def.xml

the name

def.xml

xbuild, however, uses the name

abc.def.xml

This moves the fsharpbinding to use the right resource names.