[Core] Disable xbuild by default again
We can't really enable this unless we have targets files for all project
types. XBuild will barf if you try to build a standard library project
which references a project which is of a type which does not have a
targets file. For example, xbuild will not be able to build a normal .NET
library project which references a MonoMac library project.