Write correct guid format to solution file
<ProjectGuid>...</ProjectGuid> in some Visual Studio generated F#
project files  are missing "{"..."}" in their guid in the text of the
project file. This is not generally a problem since it is a valid GUID
format. However the solution file format requires that these are
present.

Here we just add the required { } if they are not present.

Without this, bad things happen to MonoDevelop when opening
VS-generated F# projet files, see
https://github.com/fsharp/fsharpbinding/issues/42