Fix some version control issues
Fixed issue when adding files to a project. The files
were not added to version control.

To fix this, a new parameter has been added
to GetVersionInfo, which allows getting uncached
version info. The code that checks if a file can
be added to version control needs to get up to date
version information.